+++ modules/impex/core/gen/sap/commerce/toolset/impex/psi/impl/ImpExFullHeaderParameterImpl.java
@@ -34,6 +34,8 @@ import sap.commerce.toolset.impex.psi.*;
 import com.intellij.psi.PsiReference;
 import kotlin.jvm.functions.Function0;
 import sap.commerce.toolset.impex.constants.modifier.AttributeModifier;
+import sap.commerce.toolset.impex.psi.impl.ImpExFullHeaderParameterMixin.ParametersContext;
+import sap.commerce.toolset.impex.psi.impl.ImpExFullHeaderParameterMixin.ParametersContext.Parameter;
 
 public class ImpExFullHeaderParameterImpl extends ImpExFullHeaderParameterMixin implements ImpExFullHeaderParameter {